fix: duration_hours as float + timeout suggests narrower window
- Change duration_hours from int to float across all 6 tools — LLM can
  pass 0.5 (30min), 0.25 (15min) etc. if 1h still times out
- Athena timeout message now suggests fractional hours as fallback
- System prompt documents the progressive narrowing strategy

+++ tools/waf_block_fp.py
@@ -26,7 +26,7 @@ CONFIDENCE_RULES = """\
 
 
 @tool
-def investigate_block_fp(step: str = "investigate", ip: str = "", start_time: str = "", duration_hours: int = 6, hours_ago: int = None, rule_name: str = "") -> str:
+def investigate_block_fp(step: str = "investigate", ip: str = "", start_time: str = "", duration_hours: float = 6, hours_ago: float = None, rule_name: str = "") -> str:
     """Investigate whether a blocked request is a false positive, or proactively scan for suspected FPs.
 
     This tool provides evidence for human judgment — do not make definitive FP/TP verdicts
